Document code

+++ src/place.rs
@@ -1,14 +1,17 @@
 const AUTHORIZED_DISQUALIFIED_STRINGS: [&str; 2] = ["DNF", "DQ"];
 
+/// A [Place] can either be a rank denoted as an unsigned integer
+/// or a disqualification acronym ("DNF", "DQ") denoted as a String.
 #[derive(Debug, Clone)]
 pub enum Place {
-    Rank(u8),
+    Rank(u16),
     Disqualified(String),
 }
 
 impl Place {
+    /// Tries and convert a String to a [Place].
     pub fn from_string(place: &str) -> Result<Self, String> {
-        match place.parse::<u8>() {
+        match place.parse::<u16>() {
             Ok(place) => return Ok(Self::Rank(place)),
             Err(_) => {}
         };
